Update Time Periods
Update the Time period classes to use the new "strict_dates" option, which is false by default.
Add unit tests to cover each time period, and the methods to switching modes in the Calendar and Factory classes. This option will not be added to the FactoryInterface because it is not required functionality (it does not affect the data held withing classes, only restricting their input).